+++ source/StartPublicationFunction/__init__.py
@@ -1,4 +1,0 @@
-from pkg_resources import get_distribution
-
-__version__ = get_distribution('aws-data-exchange-publisher-coordinator').version
-__release_date__ = "10-may-2020"